Moved prepare from unit tests to InterpretedUtils.
